third_party/WebKit/LayoutTests/inspector/elements/styles/styles-add-new-rule-colon.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<script>

function test()
{
    if (!window.eventSender || !window.textInputController) {
        document.write("This test does not work in manual mode.");
        InspectorTest.completeTest();
        return;
    }

    WebInspector.inspectorView._showPanel("elements");
    InspectorTest.selectNodeAndWaitForStyles("inspected", step1);
    InspectorTest.addSniffer(WebInspector.UISourceCode.prototype, "addRevision", revisionAdded);

    var treeElement;
    var hasResourceChanged;

    function step1()
    {
        // Click "Add new rule".
        InspectorTest.addNewRule("foo, div#inspected, bar", step2);
    }

    function step2()
    {
        var section = WebInspector.panels.elements.sidebarPanes.styles.sections[0][2];
        var newProperty = section.addNewBlankProperty();
        newProperty.startEditing();
        textInputController.insertText("color");
        eventSender.keyDown(":");
        textInputController.insertText("maroon");
        eventSender.keyDown(";");
        InspectorTest.runAfterPendingDispatches(step3);
    }

    function step3() {
        InspectorTest.selectNodeAndWaitForStyles("other", step4);
    }

    function step4()
    {
        // Click "Add new rule".
        InspectorTest.addNewRule(null, onRuleAdded);

        function onRuleAdded()
        {
            InspectorTest.selectNodeAndWaitForStyles("inspected", step5);
        }
    }

    function step5()
    {
        InspectorTest.addResult("After adding new rule (inspected):");
        InspectorTest.dumpSelectedElementStyles(true, false, true, true);
        InspectorTest.selectNodeAndWaitForStyles("other", step6);
    }

    function step6()
    {
        InspectorTest.addResult("After adding new rule (other):");
        InspectorTest.dumpSelectedElementStyles(true, false, true);
        InspectorTest.completeTest();
    }

    function revisionAdded(revision)
    {
        InspectorTest.addResult("Revision added: " + this.displayName());
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that adding a new rule works after switching nodes.
</p>

<div id="inspected" style="font-size: 12px">Text</div>
<div id="other" style="color:red"></div>

</body>
</html>
